requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Caching==2.1.0

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Security
cryptography==41.0.7
bcrypt==4.0.1

# WebSocket support
python-socketio==5.9.0
python-engineio==4.7.1

# HTTP and utilities
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0

# Geolocation and mapping
geopy==2.4.0
folium==0.14.0
pytz==2023.3

# Data processing
openpyxl==3.1.2
pandas==2.0.3

# Production server
gunicorn==21.2.0

# Monitoring and logging
sentry-sdk[flask]==1.38.0

# Rate limiting
Flask-Limiter==3.5.0

# Environment and deployment
supervisor==4.2.5